backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==4.3.1
qdrant-client==1.7.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
python-multipart==0.0.6
langchain==0.0.352
cachetools==5.3.2